Use callback to overcome limitation of lack of HKT
We can't name a type that can take any lifetime. All we can do
is to define a generic function or type, but once it is instantiated,
its lifetime cannot be changed later.

In our case, we want to pass the reference to the environment, which
has lifetime longer than `tick` function, but shorter than 'static.
We can't name the lifetime at the moment Behavior Tree is constructed.
We need Higher Kindeded Types to describe such a type.

As a workaround, we do not pass the reference to communicate between
the environment and the behavior nodes. We use owned types, such as
a unit struct, to instruct the environment to do some actions.
`tick()` takes a function object to process such commands.

This solution is not very beautiful, but I can't find a better one
without violating ownership and borrowing model.

BTW, std::any::Any won't allow us to put arbitrary lifetime.
`Any` is always bounded by 'static.