An implementation of Symbol with a raw pointer
It reduces the symbol size to a single pointer, instead of a
pointer and usize like &str, which can help reduce the pressure
to the cache memory.

However, it has downside that it cannot be casted back to &str
or compared to a str. It basically "fogets" about being a str,
and becomes an opaque pointer.

I will leave it as a experiental branch and won't merge it.