Add tick_child_node and BehaviorNodeContainer constructor for decorators
Right now it's not possible to implement a decorator node from outside
the crate, because you cannot access Context::blackboard_map to save it.
We should allow the user to implement their own decorator nodes.

`BehaviorNodeContainer::new` can be used to wrap a BehaviorNode as a
child node outside this crate. `tick_child_node` allows the user to tick
it and save blackboard map.